+++ includes/platformer-engine/Engine/Engine.hpp
@@ -3,6 +3,7 @@
 
 #include "Render/Window.hpp"
 #include "Timer/Timer.hpp"
+#include "Physics/PhysicsSystem.hpp"
 #include "GameLevel/GameLevel.hpp"
 #include <memory>
 
@@ -72,6 +73,7 @@ namespace platformer_engine {
         bool _isRunning = false;
 
         std::unique_ptr<Window> _window = nullptr;
+        std::unique_ptr<PhysicsSystem> _physicsSystem = nullptr;
     };
 }//namespace platformer_engine