+++ includes/spic-prj-api/Component.hpp
@@ -8,32 +8,40 @@ namespace spic {
      */
     class Component {
         public:
+            Component() = default;
             /**
              * @brief Virtual destructor.
              * @spicapi
              */
             virtual ~Component() = default;
+            Component(const Component& other) = default;
+            auto operator=(const Component& other) -> Component& = default;
+            Component(Component&& other) = default;
+
+
+
+        auto operator=(Component&& other) -> Component& = default;
 
             /**
              * @brief Getter for active status.
              * @return true if active, false otherwise.
              * @spicapi
              */
-            bool Active() const { return active; }
+            [[nodiscard]] auto Active() const -> bool { return _active; }
 
             /**
              * @brief flag New active status.
              * @spicapi
              */
-            void Active(bool flag) { active = flag; }
+            void Active(bool flag) { _active = flag; }
 
         private:
             /**
              * @brief Active status.
              */
-            bool active;
+            bool _active;
     };
 
-}
+} // namespace spic
 
 #endif // COMPONENT_H_