+++ src/API/GameObject.cpp
@@ -3,75 +3,81 @@
 using namespace spic;
 
 //Creates the static instances of GameObjects
-std::map<std::string, std::shared_ptr<GameObject>> GameObject::instances;
+std::map<std::string, std::shared_ptr<GameObject>> GameObject::_instances;
 
 GameObject::GameObject(const std::string &name) {
     std::string objName = name;
-    if(instances.count(name) > 0)
+    if(_instances.count(name) > 0)
         objName += "- Copy";
 
-    this->name = objName;
+    this->_name = objName;
     auto selfptr = std::make_shared<GameObject>(*this);
-    self = selfptr;
-    instances[objName] = selfptr;
+    _self = selfptr;
+    selfptr->_self = selfptr;
+    _instances[objName] = selfptr;
 }
 
-GameObject::GameObject(const std::string &name, const std::string& tag) : tag(tag) {
+GameObject::GameObject(const std::string &name, const std::string& tag) : _tag(tag) {
     std::string objName = name;
-    if(instances.count(name) > 0)
+    if(_instances.count(name) > 0)
         objName += "- Copy";
 
-    this->name = objName;
+    this->_name = objName;
     auto selfptr = std::make_shared<GameObject>(*this);
-    self = selfptr;
-    instances[objName] = selfptr;
+    _self = selfptr;
+    selfptr->_self = selfptr;
+    _instances[objName] = selfptr;
 }
 
-bool GameObject::operator==(const spic::GameObject &other) { return name==other.name; }
+bool GameObject::operator==(const spic::GameObject &other) { return _name==other._name; }
 
-bool GameObject::operator!=(const spic::GameObject &other) { return name!=other.name; }
+bool GameObject::operator!=(const spic::GameObject &other) { return _name!=other._name; }
 
 GameObject::operator bool() { return true; } //Docs don't know what it is used for
 
 
 void GameObject::AddChild(std::shared_ptr<GameObject> child) {
-    child->parent = GameObject::Find(this->name);
-    children.emplace_back(child);
+    child->_parent = GameObject::Find(this->_name);
+    _children.emplace_back(child);
 }
 
-std::string GameObject::GetName() { return name; }
+auto GameObject::GetName() -> std::string { return _name; }
 
-std::vector<std::shared_ptr<GameObject>> GameObject::Children() { return children; }
+auto GameObject::Children() -> std::vector<std::shared_ptr<GameObject>> { return _children; }
 
-std::shared_ptr<GameObject> GameObject::Parent() { return parent; }
+auto GameObject::Parent() -> std::shared_ptr<GameObject> { return _parent; }
 
 
-std::shared_ptr<GameObject> GameObject::Find(const std::string &name) {
-    return instances[name];
+auto GameObject::Find(const std::string &name) -> std::shared_ptr<GameObject> {
+    if(_instances.count(name) > 0)
+        return _instances[name];
+
+    return nullptr;
 }
 
-std::vector<std::shared_ptr<GameObject>> GameObject::FindGameObjectsWithTag(const std::string &tag) {
+auto GameObject::FindGameObjectsWithTag(const std::string &tag) -> std::vector<std::shared_ptr<GameObject>> {
     std::vector<std::shared_ptr<GameObject>> result;
-    for(auto const& [key, val] : instances) {
-        if (val->tag == tag)
+    for(auto const& [key, val] : _instances) {
+        if (val->_tag == tag)
             result.emplace_back(val);
     }
     return result;
 }
 
-std::shared_ptr<GameObject> GameObject::FindWithTag(const std::string &tag) {
-    for(auto const& [key, val] : instances) {
-        if (val->tag == tag)
+auto GameObject::FindWithTag(const std::string &tag) -> std::shared_ptr<GameObject> {
+    for(auto const& [key, val] : _instances) {
+        if (val->_tag == tag)
             return val;
     }
+    return nullptr;
 }
 
 void GameObject::Destroy(spic::Component *obj) {
     if(obj == nullptr)
         throw std::runtime_error("Given pointer is empty or invalid");
 
-    for(auto& [name, instance] : instances) { //For every gameobject
-        for(auto& [type, cList] : instance->components ) { //For every componentType
+    for(auto& [name, instance] : _instances) { //For every gameobject
+        for(auto& [type, cList] : instance->_components ) { //For every componentType
             for(auto iter = cList.begin(); iter != cList.end(); iter++) { //For every component in list
                 if(obj == iter->get()) {
                     cList.erase(iter);
@@ -86,30 +92,37 @@ void GameObject::Destroy(std::shared_ptr<GameObject> obj) {
     if(obj == nullptr)
         throw std::runtime_error("Given pointer is empty or invalid");
 
-    for(auto& child : obj->children) {
+    for(auto& child : obj->_children) {
         Destroy(child);
     }
 
-    std::shared_ptr<GameObject> gameObject = Find(obj->name);
-    instances.erase(gameObject->name);
+    std::shared_ptr<GameObject> gameObject = Find(obj->_name);
+    _instances.erase(gameObject->_name);
     gameObject.reset();
 }
 
 
-void GameObject::Active(bool flag) { self.lock()->active = flag; }
-bool GameObject::Active() const { return self.lock()->active; }
 
-bool GameObject::IsActiveInWorld() const {
+void GameObject::Active(bool flag) { _self.lock()->_active = flag; }
+bool GameObject::Active() const { return _self.lock()->_active; }
+
+auto GameObject::IsActiveInWorld() const -> bool {
     if(!Active())
         return false;
 
-    auto par = self.lock()->parent;
+    auto par = _self.lock()->_parent;
 
     if(par != nullptr)
         return par->IsActiveInWorld();
 
     return true;
 }
+
+void GameObject::SetTransform(const spic::Transform &transform) {
+    _self.lock()->_transform = transform;
+}
+auto GameObject::GetTransform() -> Transform { return _self.lock()->_transform; }
+