+++ src/API/UIObject.cpp
@@ -3,6 +3,7 @@
 spic::UIObject::UIObject(const std::string &name, const std::string &tag, double width, double height)
                             : GameObject(name, tag), width(width), height(height) {
     auto selfptr = std::make_shared<UIObject>(*this);
-    self = selfptr;
-    instances[this->name] = selfptr;
+    _self = selfptr;
+    selfptr->_self = selfptr;
+    _instances[this->_name] = selfptr;
 }