+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Set up the project and add the first version of the code.
 - SPCH-31: Added Google Test
 - SPCH-4: Added Methods to check if mouse/keyboard inputs are being pressed, and get the mouse position
 - SPCH-31: Added Google Test 
+- SPCH-38: Added Collision Detection with direction of collision.
 
 ### Bug Fixes
 - SPCH-50: Fixed the freezing window issue